se añade boton eliminar y diseño basico en theme

+++ app/src/main/java/com/pointcheck/model/Reservation.kt
@@ -6,10 +6,9 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "reservations")
 data class Reservation(
     @PrimaryKey(autoGenerate = true)
-    val id: Long = 0,
+    val id: Int = 0,
     val userEmail: String,
     val name: String,
     val epochMillis: Long,
     val createdAt: Long = System.currentTimeMillis()
 )
-
